requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
pydantic==2.7.3
python-dotenv==1.0.1
streamlit==1.44.1
pandas==2.2.2
langchain==1.2.0
langchain-openai==1.1.6